check_map関数を移動
