+++ srcs/parser/map_parser.c
@@ -6,7 +6,7 @@
 /*   By: myokono <myokono@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 10:39:02 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/27 19:45:16 by myokono          ###   ########.fr       */
+/*   Updated: 2025/04/29 15:52:01 by myokono          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,86 +27,7 @@ static void	free_map_lines(char **lines, int height)
 	free(lines);
 }
 
-// static char **append_line(char **old_lines, int height, char *line)
-// {
-// 	char **new_lines = malloc(sizeof(char *) * (height + 2));
-// 	int i;
 
-// 	if (!new_lines)
-// 		return (NULL);
-// 	for (i = 0; i < height; i++)
-// 		new_lines[i] = old_lines[i];
-// 	new_lines[height] = line;
-// 	new_lines[height + 1] = NULL;
-// 	free(old_lines);
-// 	return new_lines;
-// }
-
-int	parse_map(t_game *game, char *filename)
-{
-	int		fd;
-	char	*line;
-	char	**map_lines;
-	char	**new_map_lines;
-	int		height;
-	int		i;
-
-	fd = open(filename, O_RDONLY);
-	if (fd < 0)
-		return (error_msg(ERR_FILE));
-	if (!parse_config(game, fd))
-		return (close(fd), 0);
-	map_lines = NULL;
-	height = 0;
-	line = NULL;
-	while (get_next_line(fd, &line) > 0)
-	{
-		if (line[0] == '\0')
-		{
-			free(line);
-			continue ;
-		}
-		new_map_lines = malloc(sizeof(char *) * (height + 2));
-		if (!new_map_lines)
-		{
-			free(line);
-			close(fd);
-			printf("here");
-			return (error_msg(ERR_MEMORY));
-		}
-		i = 0;
-		while (i < height)
-		{
-			new_map_lines[i] = map_lines[i];
-			i++;
-		}
-		new_map_lines[height] = line;
-		new_map_lines[height + 1] = NULL;
-		free(map_lines);
-		map_lines = new_map_lines;
-		height++;
-	}
-	free(line);
-	close(fd);
-	if (height == 0)
-		return (free(map_lines), error_msg(ERR_MAP));
-	if (!convert_map_data(game, map_lines, height))
-		return (free_map_lines(map_lines, height), 0);
-	if (!check_map(game))
-	{
-		free_map_lines(map_lines, height);
-		free_map(&game->map);
-		return (0);
-	}
-	if (!find_player(game))
-	{
-		free_map_lines(map_lines, height);
-		free_map(&game->map);
-		return (0);
-	}
-	free_map_lines(map_lines, height);
-	return (1);
-}
 
 int	convert_map_data(t_game *game, char **map_lines, int height)
 {
@@ -193,34 +114,70 @@ int	find_player(t_game *game)
 	return (1);
 }
 
-int	check_map(t_game *game)
-{
-	int	i;
-	int	j;
 
-	i = 0;
-	while (i < game->map.height)
+
+int	parse_map(t_game *game, char *filename)
+{
+	int		fd;
+	char	*line;
+	char	**map_lines;
+	char	**new_map_lines;
+	int		height;
+	int		i;
+
+	fd = open(filename, O_RDONLY);
+	if (fd < 0)
+		return (error_msg(ERR_FILE));
+	if (!parse_config(game, fd))
+		return (close(fd), 0);
+	map_lines = NULL;
+	height = 0;
+	line = NULL;
+	while (get_next_line(fd, &line) > 0)
 	{
-		j = 0;
-		while (j < game->map.width)
+		if (line[0] == '\0')
 		{
-			if (game->map.grid[i][j] == '0' || is_player(game->map.grid[i][j]))
-			{
-				if (i == 0 || i == game->map.height - 1 || \
-					j == 0 || j == game->map.width - 1)
-					return (error_msg(ERR_MAP));
-				if (game->map.grid[i-1][j] == ' ' || \
-					game->map.grid[i+1][j] == ' ' || \
-					game->map.grid[i][j-1] == ' ' || \
-					game->map.grid[i][j+1] == ' ')
-					return (error_msg(ERR_MAP));
-			}
-			if (game->map.grid[i][j] != '0' && game->map.grid[i][j] != '1' && \
-				game->map.grid[i][j] != ' ' && !is_player(game->map.grid[i][j]))
-				return (error_msg(ERR_MAP));
-			j++;
+			free(line);
+			continue ;
 		}
-		i++;
+		new_map_lines = malloc(sizeof(char *) * (height + 2));
+		if (!new_map_lines)
+		{
+			free(line);
+			close(fd);
+			printf("here");
+			return (error_msg(ERR_MEMORY));
+		}
+		i = 0;
+		while (i < height)
+		{
+			new_map_lines[i] = map_lines[i];
+			i++;
+		}
+		new_map_lines[height] = line;
+		new_map_lines[height + 1] = NULL;
+		free(map_lines);
+		map_lines = new_map_lines;
+		height++;
 	}
+	free(line);
+	close(fd);
+	if (height == 0)
+		return (free(map_lines), error_msg(ERR_MAP));
+	if (!convert_map_data(game, map_lines, height))
+		return (free_map_lines(map_lines, height), 0);
+	if (!check_map(game))
+	{
+		free_map_lines(map_lines, height);
+		free_map(&game->map);
+		return (0);
+	}
+	if (!find_player(game))
+	{
+		free_map_lines(map_lines, height);
+		free_map(&game->map);
+		return (0);
+	}
+	free_map_lines(map_lines, height);
 	return (1);
 }
